Rename @OAuth2Client to @RegisteredOAuth2AuthorizedClient
Fixes gh-5360
