+++ samples/boot/oauth2login/src/main/java/sample/web/OAuth2LoginController.java
@@ -17,7 +17,7 @@ package sample.web;
 
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
-import org.springframework.security.oauth2.client.annotation.OAuth2Client;
+import org.springframework.security.oauth2.client.annotation.RegisteredOAuth2AuthorizedClient;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -32,7 +32,7 @@ public class OAuth2LoginController {
 
 	@GetMapping("/")
 	public String index(Model model,
-						@OAuth2Client OAuth2AuthorizedClient authorizedClient,
+						@RegisteredOAuth2AuthorizedClient OAuth2AuthorizedClient authorizedClient,
 						@AuthenticationPrincipal OAuth2User oauth2User) {
 		model.addAttribute("userName", oauth2User.getName());
 		model.addAttribute("clientName", authorizedClient.getClientRegistration().getClientName());
